Finished button notify()
The notify portion appears to be working correctly, but I can't fully
test until the subscriber is working.  I am trying to figure out a
compromise in the subscriber that will allow me to iterate through
a list of buttons in the update method, while preserving descriptive
button names.  The end goal is to have a given functional class
inherit the subscriber class, so it has to be set up in a way,
that the functional class would be creating and naming the buttons.