Well, at least the observer works.
The good news is that the observer is working... sort of.
It recieves the notifications in the subscriber.  I still need
to figure out a way for other classes to inherit this information.

There is also another problem.  I anticipated this, but I'm having
trouble fixing it.  The button manager loop is blocking, as expected
causing the thermometer to not update.  The problem is, I'm having trouble
getting std::thread to work right.