Created TempDataManager class
Re-organizing things again.  I created a TempDataManager class
which will hold a history of n temp values, selected scale etc.
the class need to be able to register for button action as it will
directly respont to mode changes.  I was unable to figure out how
to make the TempDataManager class inherit properties of the Button
Subscriber, so for now, anything that needs a button subscription
will need the update/subscribe/unsubscribe functions and the Button
Subscriber class is essentually a dummy class.  The next step will
be to re-write the TempCtrl class so that it dependancy injects the
Button Manager class and give it a better name. Like Thermocouple
Reciever. For the record Button actions haven't been tested in the
TempDataManager yet.