Working on button logic
Buttons registered in the datamanager are monitored within the event loop.
This needs to be cleaned up a bit.  One of the operator overloads in the
button class didn't appear to be working right so it has been commented out
until it can be fixed.  Some of the logic is correct but a falling edge
following a long hold should be ignored.  Also, in refernce to the button
logic. constants should be in the common class, if they are absolutely
necessary.