Established branch for wiring pi
The eventual goal is to optimize this software and
remove dependancies including what is now the un-maintained
wiring pi.  However writing the I2C controller and the PWM
controller will be a significant amount of work by themselves.
This will also be the most portable branch in the near future.