Started replacing pointers with smart pointers
I sort of mucked the memory management together when this project
began, so I now begin the long arguous process of getting memory
management up to this decade.  I'm only parttially done.  Also,
I may switch to unique pointers for some of the threads.
Finally I probably need to make a factory class to old all this
junk and make the threads a nice neat array.  There are probably
going to be many ammended commits in this effort.  I'm commiting
now because I have no idea when I'll get back to this.